moved step4 to step5 + color step :muscle:
